Make upload endpoint tests be fully end-to-end
