+++ src/frontend/http.rs
@@ -444,17 +444,11 @@ pub async fn run_server(context: Arc<dyn SeafowlContext>, config: HttpFrontend) 
 
 #[cfg(test)]
 mod tests {
-    use arrow::array::{Int32Array, StringArray};
-    use arrow::csv::WriterBuilder;
-    use arrow::datatypes::{DataType, Field, Schema};
-    use arrow::record_batch::RecordBatch;
+
     use bytes::Bytes;
-    use datafusion::assert_batches_eq;
-    use datafusion::from_slice::FromSlice;
-    use datafusion::parquet::arrow::ArrowWriter;
+
     use itertools::Itertools;
 
-    use std::io::Cursor;
     use std::{collections::HashMap, sync::Arc};
 
     use warp::{Filter, Rejection, Reply};
@@ -464,8 +458,6 @@ mod tests {
         hyper::{header::IF_NONE_MATCH, StatusCode},
         test::request,
     };
-
-    use test_case::test_case;
 
     use crate::auth::AccessPolicy;
 
@@ -536,108 +528,6 @@ mod tests {
         "038966de9f6b9a901b20b4c6ca8b2a46009feebe031babc842d43690c0bc222b";
     const V2_ETAG: &str =
         "06d033ece6645de592db973644cf7357255f24536ff7b03c3b2ace10736f7636";
-
-    // Create a mock upload request and execute it
-    async fn mock_upload_request(
-        filename: &str,
-        file_content: Vec<u8>,
-        schema_name: &str,
-        table_name: &str,
-        schema_json: Option<&str>,
-        has_headers: Option<&str>,
-        context: Arc<dyn SeafowlContext>,
-    ) -> Response<Bytes> {
-        fn append_parts_to_body(
-            mut body: Vec<u8>,
-            boundary: &str,
-            name: &str,
-            content_type: &str,
-            filename: Option<&str>,
-            mut part: Vec<u8>,
-            end: bool,
-        ) -> Vec<u8> {
-            body.append(
-                &mut format!(
-                    "--{}\r\n\
-                Content-Disposition: form-data; name=\"{}\"{}\n\
-                Content-Type: {}\n\n",
-                    boundary,
-                    name,
-                    if let Some(filename) = filename {
-                        format!("; filename=\"{}\"", filename)
-                    } else {
-                        String::new()
-                    },
-                    content_type
-                )
-                .as_bytes()
-                .to_vec(),
-            );
-
-            body.append(&mut part);
-            if end {
-                body.append(&mut format!("--{}--", boundary).as_bytes().to_vec())
-            } else {
-                body.append(&mut "\n\n".as_bytes().to_vec());
-            };
-
-            body
-        }
-
-        let handler = filters(context, free_for_all());
-
-        let boundary = "0123456789";
-        let mut body: Vec<u8> = vec![];
-
-        if let Some(has_headers) = has_headers {
-            body = append_parts_to_body(
-                body,
-                boundary,
-                "has_header",
-                "application/json",
-                None,
-                has_headers.as_bytes().to_vec(),
-                false,
-            );
-        }
-
-        if let Some(schema_json) = schema_json {
-            body = append_parts_to_body(
-                body,
-                boundary,
-                "schema",
-                "application/json",
-                None,
-                schema_json.as_bytes().to_vec(),
-                false,
-            );
-        }
-
-        body = append_parts_to_body(
-            body,
-            boundary,
-            "data",
-            "application/octet-stream",
-            Some(filename),
-            file_content,
-            true,
-        );
-
-        request()
-            .method("POST")
-            .path(format!("/upload/{}/{}", schema_name, table_name).as_str())
-            .header("Host", "localhost:3030")
-            .header("User-Agent", "curl/7.64.1")
-            .header("Accept", "*/*")
-            .header("Content-Length", 232)
-            .header(
-                "Content-Type",
-                format!("multipart/form-data; boundary={}", boundary),
-            )
-            .body(body)
-            .reply(&handler)
-            .await
-    }
 
     #[tokio::test]
     async fn test_get_cached_hash_mismatch() {
@@ -1120,254 +1010,5 @@ SELECT
             r#"{"smallint_val":1,"integer_val":1000000,"bigint_val":10000000000,"char_val":"c","varchar_val":"varchar","text_val":"text","float_val":12.345,"real_val":12.345,"double_val":12.345678910111213,"bool_val":true,"date_val":"2022-01-01","timestamp_val":"2022-01-01 12:03:11.123456","int_array_val":[1,2,3,4,5],"text_array_val":["one","two"]}
 "#
         );
-    }
-
-    #[test_case(
-        "csv",
-        true,
-        Some(true);
-        "CSV file schema supplied w/ headers")
-    ]
-    #[test_case(
-        "csv",
-        true,
-        Some(false);
-        "CSV file schema supplied w/o headers")
-    ]
-    #[test_case(
-        "csv",
-        false,
-        Some(true);
-        "CSV file schema inferred w/ headers")
-    ]
-    #[test_case(
-        "csv",
-        false,
-        Some(false);
-        "CSV file schema inferred w/o headers")
-    ]
-    #[test_case(
-        "parquet",
-        false,
-        None;
-        "Parquet file")
-    ]
-    #[tokio::test]
-    async fn test_upload_base(
-        file_format: &str,
-        include_schema: bool,
-        add_headers: Option<bool>,
-    ) {
-        let context = in_memory_context_with_single_table().await;
-
-        let table_name = format!("{}_table", file_format);
-
-        // Prepare the schema + data (a single record batch) which we'll convert to bytes via
-        // a corresponding writer
-        let schema = Arc::new(Schema::new(vec![
-            Field::new("number", DataType::Int32, false),
-            Field::new("parity", DataType::Utf8, false),
-        ]));
-
-        // For CSV uploads we can supply the schema as another part of the multipart request, to
-        // reduce the ambiguity resulting from automatic schema inference
-        let schema_json = r#"{
-                "fields": [
-                    {
-                        "name": "number",
-                        "nullable": false,
-                        "type": {
-                            "name": "int",
-                            "bitWidth": 32,
-                            "isSigned": true
-                        }
-                    },
-                    {
-                        "name": "parity",
-                        "nullable": false,
-                        "type": {
-                            "name": "utf8"
-                        }
-                    }
-                ]
-            }"#;
-
-        let range = 0..2000;
-        let mut input_batch = RecordBatch::try_new(
-            schema.clone(),
-            vec![
-                Arc::new(Int32Array::from_slice(range.clone().collect_vec())),
-                Arc::new(StringArray::from(
-                    range
-                        .map(|number| if number % 2 == 0 { "even" } else { "odd" })
-                        .collect_vec(),
-                )),
-            ],
-        )
-        .unwrap();
-
-        // Write out the CSV/Parquet format data into mock request
-        let mut form_data: Cursor<Vec<u8>> = Default::default();
-        // drop the writer early to release the borrow.
-        if file_format == "csv" {
-            let mut writer = WriterBuilder::new()
-                .has_headers(if let Some(has_headers) = add_headers {
-                    has_headers
-                } else {
-                    true
-                })
-                .build(&mut form_data);
-            writer.write(&input_batch).unwrap();
-        } else if file_format == "parquet" {
-            let mut writer = ArrowWriter::try_new(&mut form_data, schema, None).unwrap();
-            writer.write(&input_batch).unwrap();
-            writer.close().unwrap();
-        }
-        form_data.set_position(0);
-
-        let resp = mock_upload_request(
-            format!("test.{}", file_format).as_str(),
-            form_data.into_inner(),
-            "test_upload",
-            table_name.as_str(),
-            if include_schema && file_format == "csv" {
-                Some(schema_json)
-            } else {
-                None
-            },
-            add_headers.map(|h| if h { "true" } else { "false" }),
-            context.clone(),
-        )
-        .await;
-
-        assert_eq!(resp.status(), StatusCode::OK);
-        assert_eq!(resp.body(), "done");
-
-        context.reload_schema().await;
-
-        // Verify the newly created table contents
-        let plan = context
-            .plan_query(format!("SELECT * FROM test_upload.{}", table_name).as_str())
-            .await
-            .unwrap();
-        let results = context.collect(plan).await.unwrap();
-
-        // Generate expected output from the input batch that was used in the multipart request
-        if !include_schema && add_headers == Some(false) {
-            // Rename the columns, as they will be inferred without a schema
-            input_batch = RecordBatch::try_from_iter_with_nullable(vec![
-                ("column_1", input_batch.column(0).clone(), false),
-                ("column_2", input_batch.column(1).clone(), false),
-            ])
-            .unwrap();
-        }
-        let formatted = arrow::util::pretty::pretty_format_batches(&[input_batch])
-            .unwrap()
-            .to_string();
-
-        let expected: Vec<&str> = formatted.trim().lines().collect();
-
-        assert_batches_eq!(expected, &results);
-    }
-
-    #[tokio::test]
-    async fn test_upload_to_existing_table() {
-        let context = in_memory_context_with_single_table().await;
-
-        // Prepare the schema that matches the existing table + some data
-        let schema = Arc::new(Schema::new(vec![Field::new(
-            "col_1",
-            DataType::Int32,
-            true,
-        )]));
-
-        let input_batch = RecordBatch::try_new(
-            schema.clone(),
-            vec![Arc::new(Int32Array::from(vec![2, 3]))],
-        )
-        .unwrap();
-
-        // Write out the new data into mock request for Parquet upload
-        let mut form_data: Cursor<Vec<u8>> = Default::default();
-        // drop the writer early to release the borrow.
-        {
-            let mut writer = ArrowWriter::try_new(&mut form_data, schema, None).unwrap();
-            writer.write(&input_batch).unwrap();
-            writer.close().unwrap();
-        }
-        form_data.set_position(0);
-
-        // Create a mock request and execute it
-        let resp = mock_upload_request(
-            "test.parquet",
-            form_data.into_inner(),
-            "public",
-            "test_table",
-            None,
-            None,
-            context.clone(),
-        )
-        .await;
-
-        assert_eq!(resp.status(), StatusCode::OK);
-        assert_eq!(resp.body(), "done");
-
-        context.reload_schema().await;
-
-        // Verify the newly created table contents
-        let plan = context
-            .plan_query("SELECT * FROM test_table")
-            .await
-            .unwrap();
-        let results = context.collect(plan).await.unwrap();
-
-        let expected = vec![
-            "+-------+",
-            "| col_1 |",
-            "+-------+",
-            "| 1     |",
-            "| 2     |",
-            "| 3     |",
-            "+-------+",
-        ];
-
-        assert_batches_eq!(expected, &results);
-
-        // Now try with schema that doesn't matches the existing table (re-use inout batch from before)
-        let schema = Arc::new(Schema::new(vec![Field::new(
-            "col_1",
-            DataType::Int32,
-            false,
-        )]));
-
-        let input_batch = RecordBatch::try_new(
-            schema.clone(),
-            vec![Arc::new(Int32Array::from(vec![4, 5]))],
-        )
-        .unwrap();
-
-        // Write out the new data into mock request for Parquet upload
-        let mut form_data: Cursor<Vec<u8>> = Default::default();
-        // drop the writer early to release the borrow.
-        {
-            let mut writer = ArrowWriter::try_new(&mut form_data, schema, None).unwrap();
-            writer.write(&input_batch).unwrap();
-            writer.close().unwrap();
-        }
-        form_data.set_position(0);
-
-        let resp = mock_upload_request(
-            "test.parquet",
-            form_data.into_inner(),
-            "public",
-            "test_table",
-            None,
-            None,
-            context,
-        )
-        .await;
-
-        assert_eq!(resp.status(), StatusCode::BAD_REQUEST);
-        assert_eq!(resp.body(), "Execution error: The table public.test_table already exists but has a different schema than the one provided.");
     }
 }
