+++ tests/end_to_end.rs
@@ -1,3 +1,5 @@
+mod http;
+
 use std::env;
 use std::sync::Arc;
